backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0